Add basic temp sweep code
