+++ src/mpmorph/jobs/lammps/templates/basic_constant_temp.lammps
